Ajuste tipo de dado ID

+++ src/main/resources/db/migration/V1__CREATE_TABLE_USUARIO.sql
@@ -1,5 +1,5 @@
 CREATE TABLE usuario (
-	id int8 NOT NULL,
+	id serial NOT NULL,
 	nome varchar(255) NOT NULL,
 	idade integer NOT NULL,
 	CONSTRAINT usuario_pkey PRIMARY KEY (id)
